# backend/docs/swagger.yaml
basePath: /
definitions:
  calculator.BinaryOperationRequest:
    properties:
      operand_a:
        example: 10.5
        type: number
      operand_b:
        example: 3.2
        type: number
    type: object
  calculator.CalculateResponse:
    properties:
      result:
        example: 13.7
        type: number
    type: object
  calculator.ErrorResponse:
    properties:
      error:
        example: division by zero
        type: string
    type: object
  calculator.UnaryOperationRequest:
    properties:
      operand:
        example: 25
        type: number
    type: object
info:
  contact: {}
  description: REST API for arithmetic and advanced calculator operations.
  title: FS Calculator API
  version: "1.0"
paths:
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - Health
  /v1/operations/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Operands
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/calculator.BinaryOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calculator.CalculateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/calculator.ErrorResponse'
      summary: Add two numbers
      tags:
      - Operations
  /v1/operations/divide:
    post:
      consumes:
      - application/json
      parameters:
      - description: Operands
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/calculator.BinaryOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calculator.CalculateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/calculator.ErrorResponse'
      summary: Divide two numbers
      tags:
      - Operations
  /v1/operations/multiply:
    post:
      consumes:
      - application/json
      parameters:
      - description: Operands
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/calculator.BinaryOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calculator.CalculateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/calculator.ErrorResponse'
      summary: Multiply two numbers
      tags:
      - Operations
  /v1/operations/percentage:
    post:
      consumes:
      - application/json
      parameters:
      - description: Operands
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/calculator.BinaryOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calculator.CalculateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/calculator.ErrorResponse'
      summary: Calculate percentage
      tags:
      - Operations
  /v1/operations/power:
    post:
      consumes:
      - application/json
      parameters:
      - description: Operands
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/calculator.BinaryOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calculator.CalculateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/calculator.ErrorResponse'
      summary: Raise a number to a power
      tags:
      - Operations
  /v1/operations/sqrt:
    post:
      consumes:
      - application/json
      parameters:
      - description: Operand
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/calculator.UnaryOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calculator.CalculateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/calculator.ErrorResponse'
      summary: Calculate a square root
      tags:
      - Operations
  /v1/operations/subtract:
    post:
      consumes:
      - application/json
      parameters:
      - description: Operands
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/calculator.BinaryOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/calculator.CalculateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/calculator.ErrorResponse'
      summary: Subtract two numbers
      tags:
      - Operations
swagger: "2.0"